提交主机clone 接口
